Register EventBus in a BaseActivity
