+++ brewchat/src/main/java/com/example/brewchat/activities/BaseActivity.java
@@ -1,0 +1,21 @@
+package com.example.brewchat.activities;
+
+import android.support.v7.app.AppCompatActivity;
+
+import de.greenrobot.event.EventBus;
+
+public class BaseActivity extends AppCompatActivity {
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        EventBus.getDefault().register(this);
+    }
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+        EventBus.getDefault().unregister(this);
+    }
+
+}
